pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanogcg"
version = "0.3.0"
authors = [
    { name="Justin Wang", email="justin@grayswan.ai" },
]
description = "A lightweight implementation of the GCG algorithm"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20.0",
    "protobuf",
    "scipy>=1.7.0",
    "sentencepiece>=0.2.0",
    "torch>=2.0",
    "tqdm>=4.60.0",
    "transformers>=4.4,<=4.47.1", # FIXME: Use `transformers.Cache` class for `past_key_values`
]

[project.optional-dependencies]
constrained = [
    "faiss-cpu>=1.7.0",  # Use faiss-gpu for GPU acceleration
]

[tool.setuptools]
packages = ["nanogcg"]

[project.urls]
Homepage = "https://github.com/GraySwanAI/nanoGCG"
Issues = "https://github.com/GraySwanAI/nanoGCG/issues"
